update: 데이터 doc에 x, y 필드 추가
