+++ src/components/CardList.jsx
@@ -18,7 +18,7 @@ function CardList() {
   }
 
   // console.log(data);
-  data.forEach((item) => console.log(`위도 : ${item.x} 경도: ${item.y}`));
+  // data.forEach((item) => console.log(`위도 : ${item.x} 경도: ${item.y}`));
 
   return (
     <StCardList>
